[feat/#1178] 기타 코드 수정
