+++ app/src/main/java/org/sopt/official/feature/auth/AuthSideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.official.feature.auth
+
+sealed class AuthSideEffect {
+    data class ShowToast(val message: String) : AuthSideEffect()
+}
