+++ app/src/main/java/org/sopt/official/feature/auth/feature/authmain/AuthMainViewModel.kt
@@ -25,13 +25,10 @@
 package org.sopt.official.feature.auth.feature.authmain
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.launch
-import org.sopt.official.domain.auth.model.Auth
 import org.sopt.official.domain.auth.repository.AuthRepository
 import javax.inject.Inject
 
@@ -40,27 +37,9 @@ class AuthMainViewModel @Inject constructor(
     private val authRepository: AuthRepository,
 ) : ViewModel() {
 
+    // TODO: 안쓰는코드 삭제하기
+
     private val _sideEffect = MutableSharedFlow<AuthMainSideEffect>()
     val sideEffect: SharedFlow<AuthMainSideEffect> = _sideEffect.asSharedFlow()
 
-    fun signIn(token: String) {
-        // TODO: 실제 token 넣기 by leeeyubin
-        viewModelScope.launch {
-            authRepository.signIn(
-                Auth(
-                    token = "codecodecodecodecode",
-                    authPlatform = GOOGLE,
-                )
-            ).onSuccess {
-                //TODO: 홈 화면으로 이동
-                _sideEffect.emit(AuthMainSideEffect.ShowToast("성공"))
-            }.onFailure {
-                _sideEffect.emit(AuthMainSideEffect.ShowToast("실패"))
-            }
-        }
-    }
-
-    companion object {
-        private const val GOOGLE = "GOOGLE"
-    }
 }
